requirements.txt
tqdm
cycler==0.10.0
decorator==4.4.0
h5py==2.9.0
imageio==2.5.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
matplotlib==3.0.3
mlxtend==0.16.0
networkx==2.3
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.2.1
six==1.12.0
